requirements.txt
# Core Dependencies
fastapi==0.115.12
uvicorn==0.34.2
python-dotenv==1.1.0
pydantic==2.11.3
pydantic-settings==2.9.1
 
# LangChain and Related
langchain==0.3.24
langchain-community==0.3.23
langchain-core==0.3.56
langchain-openai==0.3.14
langchain-text-splitters==0.3.8
langchain-experimental==0.3.0
langgraph==0.2.20
python-multipart==0.0.9
 
# Vector Store and Database
pymilvus==2.5.8
redis==5.3.0
langchain-milvus==0.2.0
 
# LLM and Embeddings
openai==1.76.0
sentence-transformers==4.1.0
transformers==4.51.3
torch==2.6.0
 
# Utilities
aiofiles==24.1.0
tiktoken==0.9.0

#gunicorn
gunicorn==23.0.0